disable/enable diplomacy action buttons stub & fix relations float format

+++ src/gui/gui_common_elements.hpp
@@ -1457,7 +1457,7 @@ public:
 	std::string get_text(sys::state& state) noexcept override {
 		auto rel = state.world.get_diplomatic_relation_by_diplomatic_pair(nation_id, state.local_player_nation);
 		auto fat_rel = dcon::fatten(state.world, rel);
-		return std::to_string(fat_rel.get_value());
+		return std::to_string(int32_t(fat_rel.get_value()));
 	}
 };
